Split tests into two files
Import steps parametrization affects
all tests in the file, but we
only want it for TestExcelReportA
